app/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
mysqlclient==2.2.6
python-dotenv==1.0.1